servo: Merge #19077 - changed f64 to u64 for navigation start timing until it had to be float (from avadacatavra:floats-are-the-actual-worst); r=jdm
<!-- Please describe your changes on the following line: -->

---
<!-- Thank you for contributing to Servo! Please replace each `[ ]` by `[X]` when the step is complete, and replace `__` with appropriate data: -->
- [X] `./mach build -d` does not report any errors
- [X] `./mach test-tidy` does not report any errors
- [X] These changes fix #18997 (github issue number if applicable).

<!-- Either: -->
- [ ] There are tests for these changes OR
- [ ] These changes do not require tests because _____

<!-- Also, please make sure that "Allow edits from maintainers" checkbox is checked, so that we can help you if you get stuck somewhere along the way.-->

<!-- Pull requests that do not address these steps are welcome, but they will require additional verification as part of the review process. -->

Source-Repo: https://github.com/servo/servo
Source-Revision: fa319170ebb34afcdfc120b7c3e47fe5b1c21210

UltraBlame original commit: babdffe6ee1d37cf35ca57a0b11f1160c5537fae

+++ servo/components/compositing/compositor.rs
@@ -1311,7 +1311,7 @@ impl<Window: WindowMethods> IOCompositor<Window> {
         
         
         if !self.pending_paint_metrics.is_empty() {
-            let paint_time = precise_time_ns() as f64;
+            let paint_time = precise_time_ns();
             let mut to_remove = Vec::new();
             
             for (id, pending_epoch) in &self.pending_paint_metrics {
